Disable the line-length check.
There are occasions where following this rule forces you to make the
code less readable. This is particularly the case for tests where method
names form the test descriptions. Keeping test descriptions under a
certain length can force them to become cryptic.

In line with the intro to the styleguide, this therefore isn't something
that's appropriate to enforce automatically.