Consider all changed files in changed_files method
Remove the `.select {}` method that restricted the `.changed_files` method to
".rb" files only.

I know that my Gemfile contains violations of the GOV.UK Styleguide codified in
the included Cops. I was surprised that `govuk-lint-ruby` didn't report on
these violations when I used the `--diff` option, even though I'd made a change
to the Gemfile in my branch.

It doesn't feel right to me that `govuk-lint-ruby --diff` is restricted to
".rb" files when `govuk-lint-ruby` isn't.