Respect local .rubocop.yml
Projects should be allowed to override the Ruby styleguide. This PR
makes govuk-lint respect any local .rubocop.yml file.

When a .rubocop.yml file is found in the current directory, we add that
config to the inherits_from in govuk-lint's original config (we need a
temporary file for this).

This means that the rules in the local file take precedence over the
ones in this repo.

The immediate use case is to start experimenting with enabling a bunch
of useful cops that are currently disabled. I'm also looking into custom
GOV.UK cops that check things like not having sessions enabled.

Fixes #14