Enable all the lint checks
Lint checks are the most useful feature of Rubocop. They check for
signs code is actually broken, not just of it's formatted correctly.

Cops like `UnreachableCode` and `UselessAssignment` point to bugs in
most cases.

I believe these cops were perhaps never meant to be disabled. There
will possibly be cops we do not want  to enable, but those will
probably be fewer than the ones we like.  Therefore this commit enables
them _all_.