Pin to a version of rubocop we know likes our code
The latest 0.36.0 release of rubocop has new cops that mean our linter code fails to pass the checks.  It also renames a bunch of existing cops which means some of our config no longer applies.  We should eventually fix this, but it'll require extra effort to evaluate the new cops (and changes to defaults in existing cops) against our styleguide and see what should be enabled, disabled, or have their default settings changed.