Update cops for Ruby to disable IfUnlessModifier
I think there are two cases when forcing single line if/unless blocks to
be re-written as one line statement with the conditional statement at
the end of the line makes it less readable.

When we have sequence of if/unless statements, some with multiple
lines within the if statement and some with a single line, forcing the
single line statements to re-written makes it a little harder to follow
the branching logic. For example:

if some_condition
  do_this
  and_this
end

just_do_this if another_condition

if last_condition
  do_this
  and_also_do_this
end

In this scenario, I would prefer to keep all the branches in the same
format, even though there is only one line in the block of the middle if 
statement, for example:

if some_condition
  do_this
  and_this
end

if another_condition
  just_do_this
end

if last_condition
  do_this
  and_also_do_this
end

The second scenario, when we have long lines of code within the a single 
line of a block inside an if statement, pushing that out to the end 
statement, I think makes it less noticeable. For example

if some_condition
  do_a_really_long_method_name_or_expression_that_takes_a_lot_of_view_space
end

verses:

do_a_really_long_method_name_or_expression_that_takes_a_lot_of_view_space if some_condition