Add Javascript linting via eslint
Note; The rules in the lint file are very much a placeholder to test
the linting process. Deciding what linting rules we'd like to apply
can come after we've got the mechanism to do it in place.

`eslint` was chosen because it is a more modern version of `jshint`,
which superceded `jslint`, and is becoming the most common JS linter.
It provides more fine-grained options than `jshint`, in particular
around ES5/6 features. It also seems to have better documentation
and support.

Our wrapper around `eslint` takes single path-or-file argument. It
doesn't support multiple target paths - or excluding sub paths, which
is something we're likely to need to do, as it's common for apps/repos
to have vendor code along-side our JS.

By default our linger it ignores `.min.js` as it's a safe bet it's not JS we have
written, or at leasr is not written as part of this repo, and is most
likely 3rd part JS - eg, jquery plugins, etc.

The linting rules are pulled from a `eslintrc` file, as this
- mirrors the approach of the SCSS and Ruby linters in this gem
- is the standard way to configure `eslint`
- makes it easy for non-ruby code to copy/programatically consume
  this file, eg sub-moduling it into a node project.