Switch to .jshintrc file from yml based config
This is more JS 'native', and how the jshint options are commonly
stored with jshint, the native bin and tooling like `grunt` plugins.

Using is makes it closer to normal world usage, easier to read and
easier to edit if you're already familiar with jshint, and without
abstraction around it.