Ruby style - constants should be frozen
By default constants in Ruby an be reassigned and mutated. Object#freeze
can be called on "CONSTANTS" to prevent any further modification.

It's also worth noting that  Ruby 2.1 includes a memory optimisation for
frozen strings whereby the same ummutable string object is always
returned, so as well as this being good practice it may also save
memory[0]

[0] - http://josephyi.com/freeze/