Allow local file .rubocop.yml override Excludes
Given how rubocop resolves the `inherit_from` and config files,
the different files will override in order the other configs hash
sub-values. Finally the root file has priority.

Because that, `AllCops/Exclude` in the main file `all.yml` will override
any optional `AppCops/Exclude` from local file passed by the user.

The solution is split the `AppCops/Exclude` and add it as an additional
file in the `inherit_from` list of files, so that the local file can override
the value if desired.

Tests has been added to ensure that the local file is the last one added
and that the root file does not include the `AllCops/Exclude` entry.

[1] https://github.com/bbatsov/rubocop/blob/49598a035cd39a75ec91466a4e1a79c7e4f0f5f9/lib/rubocop/config_loader_resolver.rb#L24