Remove PlaceholderInExtend linter from SCSS
This is a nice linter. It suggests the use of a placeholder selector rather
than a class selector, as it would generate less code [1]. However, since we use
Twitter Bootstrap as a baseline and Twitter Bootstrap does not support
placeholder selectors (due to issues with media queries [2]), this linter raises
errors that we cannot fix.

By removing this linter we suppress errors we couldn't otherwise fix.

[1] https://github.com/brigade/scss-lint/blob/master/lib/scss_lint/linter/README.md#placeholderinextend
[2] https://github.com/twbs/bootstrap-sass/issues/721