Disable Style/SafeNavigation cop
Govuk-lint now uses a newer version of Rubocop (PR #66) which includes the
Style/SafeNavigation cop and enables it by default.

This cop enforces the use of Ruby's safe navigation operator (&.).

This is a relatively new Ruby feature and is not mentioned in the Style
guide, so this commit disables it by default.