Relax constraint on rubocop from v0.43.x to v0.49.x
The changes are mostly bug fixes or new features, so I don't think
there's too much risk in relaxing the constraint to allow the latest
version (v0.49.1) of `rubocop`.

My specific motivation is the relaxation of the "Style/VariableNumber"
rule in this commit [1] which I believe was first released in v0.44.0.

[1]: bbatsov/rubocop@833aad6ccb35a16c6196830e936f4ef34384590b