Don't complain about block length in tests
Many test frameworks use blocks for the entire test case in a file as well as for individual tests within the file.  It's not unusual for a test case to be several lines longer than the upper limit of the BlockLength cop, and there's very little we can do to avoid that so there's no point running this cop on our tests.