Metrics/BlockLength for specs in sub-dirs
On GOV.UK PaaS we have a monorepo with multiple sub-directories that contain
their own `spec` directories for tests. Change the exclude pattern to allow
for this, which will also satisfy the original behaviour of a single `spec`
directory in the repository root. The string has to be quoted otherwise the
YAML parser will complain about the asterisks.