Enforce compact style naming of class and module children
This came up in a discussion we had on Platform Health during review of
a PR
(https://github.com/alphagov/local-links-manager/pull/421#discussion_r2776071940).

I prefer this style of naming classes as I think it looks cleaner and
avoids unnecessary indentation and module lines at the top and bottom of
the file.

The downside to enabling this is that classes defined in this way can't
be loaded until the parent module has already been defined. In practice,
with Rails autoloading, this means that each module would need its own
file `module.rb` containing just `module ModuleName; end`.

There are both styles currently present in the GOV.UK code base:
- https://github.com/alphagov/local-links-manager/blob/be4ad1f3bc0dee0d73f50c0819c89cb0af8c0584/lib/local-links-manager/import/links.rb#L3-L5
- https://github.com/alphagov/publishing-api/blob/fbd0e07dc3aaec156b8361d5e157ec1775ff6550/lib/link_expansion/link_reference.rb#L1

We should decide on a style and enforce it using govuk-lint. Personally,
I prefer the compact style but if the majority would prefer the nested
style, we can enforce that instead.