Fix GC race condition for temp config file
Previously we merged a bunch of YML files into a single, temporary
config file, using the ConfigFile class, which did not retain any
reference to the instance of Tempfile it created. Depending on when
Ruby's Garbage Collector runs, the dangling Tempfile instance could
have been deleted before it could be read by the RuboCop CLI itself.

This modifies the ConfigFile class and the way it's called to ensure
a reference to the Tempfile instance is maintained until the RuboCop
CLI has finished running, at which point it can be safely GC'd.