Move Rails default disabling to inherited file
This fixes an issue where Rails linting could not be switched on by
local .rubocop.yml files. This gem employs a slightly weird system where
a global rubocop config is used as the default and the local file is set
as something that this global file inherits from. This means it oddly
has higher precedence than local config files.