Bump version to 4.0.0
This is a MAJOR release due to the significant number of corrections
that will be required to many repos. We intend to script auto-correcting
the new linting issues, so that the PRs to upgrade to this new version
are ready to merge in and won't lead to sporadic corrections in future.